+++ internal/secretstore/keyring_nonlinux.go
@@ -1,0 +1,15 @@
+//go:build !linux
+
+package secretstore
+
+import "github.com/byteness/keyring"
+
+func openSystemKeyring() (keyring.Keyring, error) {
+	return keyring.Open(keyring.Config{
+		ServiceName: service,
+		AllowedBackends: []keyring.BackendType{
+			keyring.KeychainBackend,
+			keyring.WinCredBackend,
+		},
+	})
+}